+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Release History
 
-## 1.0.0-preview.4 (Unreleased)
+## 1.0.0-preview.5 (Unreleased)
+
+
+## 1.0.0-preview.4 (2020-06-09)
+
+### Added
+
+- Referencing `Azure.Core.Experimental` which brings new spatial types and custom serializers.
+- Added `SearchClientBuilderExtensions` to integrate with ASP.NET Core.
+- Added `SearchModelFactory` to mock output model types.
 
 ### Breaking Changes
 
@@ -12,8 +21,12 @@
 - Renamed `AnalyzeTextOptions.Tokenizer` to `AnalyzeTextOptions.TokenizerName`.
 - Renamed `CustomAnalyzer.Tokenizer` to `CustomAnalyzer.TokenizerName`.
 - Renamed `SearchIndexerDataSource` to `SearchIndexerDataSourceConnection`.
+- Renamed `Autocompletion` to `AutocompleteItem`.
 - Renamed methods on `SearchIndexerClient` matching "\*DataSource" to "\*DataSourceConnection".
 - Split `SearchServiceClient` into `SearchIndexClient` for managing indexes, and `SearchIndexerClient` for managing indexers, both of which are now in `Azure.Search.Documents.Indexes`.
+- `SearchClient.IndexDocuments` now throws an `AggregateException` wrapping all the `RequestFailedException`s in the batch.
+- Removed `dynamic` support from `SearchDocument` for the time being.
+
 
 ## 1.0.0-preview.3 (2020-05-05)